requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Machine learning models
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

# Deep learning (optional)
torch>=2.0.0
transformers>=4.30.0

# Data processing
python-dateutil>=2.8.2
pytz>=2023.3

# Database
sqlalchemy>=2.0.0
# sqlite3 is built-in to Python, no need to install

# API clients
requests>=2.31.0
httpx>=0.24.0

# LLM/Event processing
openai>=1.0.0  # For event classification
langchain>=0.1.0  # Optional for structured extraction

# Utilities
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# API framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Portfolio optimization
cvxpy>=1.3.0
cvxopt>=1.3.0

# Visualization (for backtesting)
matplotlib>=3.7.0
seaborn>=0.12.0

